A workable approach to zipped references
Generic code that produces lifetime-bound references into a table
entry can use Refs and MutRefs as adapter types. They are parameterized
with a lifetime and a ZippedPtrs type providing access to the values.

Instances of Refs and MutRefs can be transformed into plain references
or reference tuples in less abstract client code that instantiates either
LonePtr or PairPtrs as the type parameter. Use the suggestively named
methods `into_concrete` for that.